requirements.txt
aiogram==3.13.1
yandex-music==2.2.0
python-dotenv==1.0.0
flask==3.0.0
requests==2.31.0
